+++ apis/vara-eth/src/api/index.ts
@@ -1,3 +1,4 @@
 export { InjectedTx, InjectedTxPromise, Injected } from './injected/index.js';
 
-export * from './api.js';
+export type { VaraEthApi } from './api.js';
+export * from './factory.js';