+++ apis/vara-eth/src/signer/adapters/index.ts
@@ -1,1 +1,2 @@
 export * from './wallet-client.js';
+export * from './dynamic.js';