+++ idea/vara-eth/frontend/src/app/providers/index.ts
@@ -1,4 +1,3 @@
-import { useVaraEthApi } from './vara-eth-api';
 import { withProviders } from './with-providers';
 
-export { withProviders, useVaraEthApi };
+export { withProviders };