+++ idea/vara-eth/frontend/src/app/providers/vara-eth-api/index.ts
@@ -1,2 +1,0 @@
-export { VaraEthApiProvider } from './provider';
-export { useVaraEthApi } from './use-vara-eth-api';