+++ idea/vara-eth/frontend/src/app/providers/vara-eth-api/use-vara-eth-api.ts
@@ -1,5 +1,0 @@
-import { useContext } from 'react';
-
-import { ApiContext } from './context';
-
-export const useVaraEthApi = () => useContext(ApiContext);